Remove usage of _filter_files
According to the logs this function is deprecated and also filtering of
files can have an impact to some projects. By default it's not possible
to add dot directories, dot files or the folder templates to the
documentation.

INFO    -  DeprecationWarning: warning_filter doesn't do anything since MkDocs 1.2 and will be removed soon. All messages on the `mkdocs` logger get counted automatically.
             File "xxx/.venv/lib/python3.9/site-packages/mkdocs_multirepo_plugin/util.py", line 10, in <module>
               from mkdocs.utils import warning_filter
             File "xxx/.venv/lib/python3.9/site-packages/mkdocs/utils/__init__.py", line 453, in __getattr__
               warnings.warn(