catch all failed exit codes
Typically on UNIX systems a non-zero code is a failed exit code. The
code here doesn't catch anything except 1 but 127 is possible when it
cannot find the scripts as seen in issue #122. Higher numbers are
possible when a signal interrupts the execution. Switch to treating all
non-zero codes as an error.