convert scripts to being package resources
The code assumes that where this package is installed will result in the
scripts living in bash's PATH. If the user is using a virtualenv and
uses a shell other than bash then it won't be in the PATH and it will
fail. This uses the functionality that Python 3.9 added in importlib to
create a temp file with a resource included in the package to run the
script. For older versions of Python the backport package is installed
and used. fixes #122